# api_documentation.yaml
openapi: 3.0.3
info:
  title: API Documentation
  description: Auto-generated API documentation
  version: 1.0.0
paths:
  /vets/{vetId}/visits:
    get:
      summary: Endpoint to read visits by vet ID
      description: Retrieves all visits associated with a specific vet.
      operationId: readByVet
      tags:
      - Visits
      responses:
        '200':
          description: Success - Returns list of visits for the given vet
        '400':
          description: Bad request - Invalid vetId (less than 1)
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      parameters:
      - name: vetId
        in: path
        required: true
        schema:
          type: integer
        description: Unique identifier of the vet (must be >= 1)
  /owners/*/pets/{petId}/visits/{visitId}:
    delete:
      summary: Delete a visit by ID
      description: Deletes a visit associated with a pet by their IDs.
      operationId: deleteVisit
      tags:
      - Visit
      responses:
        '204':
          description: Visit deleted successfully
        '400':
          description: Invalid ID provided
        '404':
          description: Visit not found
        '500':
          description: Internal server error
      parameters:
      - name: petId
        in: path
        required: true
        schema:
          type: integer
        description: ID of the pet
      - name: visitId
        in: path
        required: true
        schema:
          type: integer
        description: ID of the visit to delete
  /pets/visits:
    get:
      summary: Retrieve visits by pet IDs
      description: Fetches all visit records associated with the provided list of
        pet IDs.
      operationId: getVisitsByPetIds
      tags:
      - Visits
      responses:
        '200':
          description: Success
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      parameters:
      - name: petId
        in: query
        required: true
        schema:
          type: array
          items:
            type: integer
        description: List of pet IDs to query
tags:
- name: Visit
- name: Visits